started adding more useful funcs from pathnames (e.g. probe-path)
This is starting to wander off point though, originally this was meant
to be only about paths but it's becoming obvious that pathnames are
still better & it's mainly because of the interaction with the rest of
the standard library, uiop & cl-fad.

On one hand I wonder if I should try and build a sane api on top of
uiop's functions, on the other I wonder if I just flesh out the
features and see how it feels.

Another weird bit is I want a path that makes sense for all platforms,
something I can use on windows & posix and get decent results.
I wonder if I could make a path type like that using this system.
I could then provide a pile of platform specific functions for getting
useful paths in that format.